+++ Chapter#07/FavoritesList.py
