+++ Chapter#07/PositionalList.py
@@ -1,87 +1,36 @@
-import random
-
-class _DoublyLinkedBase:
-    """ A base class providing a doubly linked list representation."""
-    class Node:
-        """Lightweight, nonpublic class for storing a doubly linked node."""
-        __slots__ = '_element' , '_prev' , '_next' # streamline memory
-
-        def __init__(self, element, prev, next): # initialize node’s fields
-            self._element = element # user’s element
-            self._prev = prev # previous node reference
-            self._next = next # next node reference
-
-    ## DoublyLinkedBase
-    def __init__(self):
-        """ Create an empty list."""
-        self._header = self.Node(None, None, None)
-        self._trailer = self.Node(None, None, None)
-        self._header._next = self._trailer # trailer is after header
-        self._trailer._prev = self._header # header is before trailer
-
-        self._size = 0 # number of elements
-
-    def __len__(self):
-        """ Return the number of elements in the list. """
-        return self._size
-
-    def is_empty(self):
-        """Return True if list is empty."""
-        return self._size == 0
-
-    def _insert_between(self, e, predecessor, successor):
-        """ Add element e between two existing nodes and return new node."""
-        newest = self. Node(e, predecessor, successor) # linked to neighbors
-        predecessor._next = newest
-        successor._prev = newest
-        self._size += 1
-        return newest
-
-    def _delete_node(self, node):
-        """ Delete nonsentinel node from the list and return its element."""
-        predecessor = node._prev
-        successor = node._next
-
-        predecessor._next = successor
-        successor._prev = predecessor
-        self._size -= 1
-
-        element = node._element # record deleted element
-        node._prev = node._next = node._element = None # deprecate node
-        return element # return deleted element
+from _DoublyLinkedBase import _DoublyLinkedBase
 
 class PositionalList(_DoublyLinkedBase):
-    """A sequential container of elements allowing positional access."""
+    """ A sequential container of elements allowing positional access."""
     #-------------------------- nested Position class --------------------------
     class Position:
         """ An abstraction representing the location of a single element."""
         def __init__(self, container, node):
-            """ Constructor should not be invoked by user. """
+            """ Constructor should not be invoked by user."""
             self._container = container
             self._node = node
+
         def element(self):
             """ Return the element stored at this Position."""
             return self._node._element
+
         def __eq__(self, other):
-            """ Return True if other is a Position representing the same location. """
+            """ Return True if other is a Position representing the same location."""
             return type(other) is type(self) and other._node is self._node
 
         def __ne__(self, other):
-            """Return True if other does not represent the same location."""
+            """ Return True if other does not represent the same location."""
             return not (self == other) # opposite of eq
-    #--------------------------End of nested Position class --------------------------
-        
+    #-------------------------- End of nested Position class --------------------
     #------------------------------- utility method -------------------------------
     def _validate(self, p):
         """ Return position s node, or raise appropriate error if invalid."""
         if not isinstance(p, self.Position):
-            raise TypeError('p must be proper Position type')
-        
+            raise TypeError("p must be proper Position type")
         if p._container is not self:
-            raise ValueError('p does not belong to this container')
+            raise ValueError("p does not belong to this container")
         if p._node._next is None: # convention for deprecated nodes
-            raise ValueError('p is no longer valid')
-        
+            raise ValueError("p is no longer valid")
         return p._node
 
     #------------------------------- utility method -------------------------------
@@ -112,7 +61,7 @@ class PositionalList(_DoublyLinkedBase):
         return self._make_position(node._next)
 
     def __iter__(self):
-        """ Generate a forward iteration of the elements of the list. """
+        """ Generate a forward iteration of the elements of the list."""
         cursor = self.first( )
         while cursor is not None:
             yield cursor.element( )
@@ -121,13 +70,13 @@ class PositionalList(_DoublyLinkedBase):
     #------------------------------- mutators -------------------------------
     # override inherited version to return Position, rather than Node
     def _insert_between(self, e, predecessor, successor):
-        """ Add element between existing nodes and return new Position. """
+        """ Add element between existing nodes and return new Position."""
         node = super()._insert_between(e, predecessor, successor)
         return self._make_position(node)
 
     def add_first(self, e):
-        """ Insert element e at the front of the list and return new Position."""
-        return self._insert_between(e,self._header, self._header._next)
+        """" Insert element e at the front of the list and return new Position."""
+        return self._insert_between(e, self._header, self._header._next)
 
     def add_last(self, e):
         """ Insert element e at the back of the list and return new Position."""
@@ -144,20 +93,36 @@ class PositionalList(_DoublyLinkedBase):
         return self._insert_between(e, original, original._next)
 
     def delete(self, p):
-        """ Remove and return the element at Position p. """
+        """ Remove and return the element at Position p."""
         original = self._validate(p)
         return self._delete_node(original) # inherited method returns element
 
     def replace(self, p, e):
         """ Replace the element at Position p with e.
-            Return the element formerly at Position p. """
+            Return the element formerly at Position p.
+        """
         original = self._validate(p)
         old_value = original._element # temporarily store old element
         original._element = e # replace with new element
         return old_value # return the old element value
 
+    def __max__(self):
+        first_position = first()
+        max_element = first_position.element
+        
+        return max_element
 
 if __name__ == '__main__':
-    L = PositionalList()
-    position = L.add_first(14)
-    print(position.element())
+    D = PositionalList()
+    D.add_first(1)
+    D.add_first(2)
+    D.add_first(3)
+    D.add_first(4)
+    D.add_last(5)
+    D.add_last(6)
+    D.add_last(7)
+
+    for e in D:
+        print(e)
+
+    print(max(D))
